linking login to user_id

+++ blog/main.py
@@ -4,7 +4,7 @@ from fastapi.encoders import jsonable_encoder
 from sqlalchemy.orm import Session, joinedload
 from blog import models, schemas, auth, database, oauth2
 from blog.database import engine, SessionLocal, get_db
-from blog.schemas import BlogSchema, ShowBlog, User
+from blog.schemas import BlogSchema, User
 from typing import List
 from sqlalchemy.ext.declarative import declarative_base
 
@@ -34,11 +34,14 @@ def show_all_users(db: Session= Depends(get_db)):
     return jsonable_encoder(users)
 
 
-@app.post('/blog', tags=['blogs'], response_model=schemas.ShowBlog, status_code=status.HTTP_201_CREATED)
-def create_blog(request: schemas.BlogSchema, db: Session = Depends(get_db), get_current_user:schemas.User = Depends(oauth2.get_current_user)):
+@app.post('/blog', tags=['blogs'], response_model=schemas.BlogSchema, status_code=status.HTTP_201_CREATED)
+def create_blog(request: schemas.BlogSchema, 
+                db: Session = Depends(get_db), 
+                current_user:models.User = Depends(oauth2.get_current_user),
+                ):
     new_blog = models.Blog(title=request.title, 
                            body=request.body, 
-                           user_id=request.user_id)
+                           user_id=current_user.id)
     
     db.add(new_blog)
     db.commit()
@@ -46,26 +49,36 @@ def create_blog(request: schemas.BlogSchema, db: Session = Depends(get_db), get_
     return new_blog
 
 
-@app.get('/blogs', tags=['blogs'], response_model=List[schemas.ShowBlog], status_code=status.HTTP_202_ACCEPTED)
-def show_all_blogs(db: Session= Depends(get_db), get_current_user:schemas.User = Depends(oauth2.get_current_user)):
+@app.get('/blogs', tags=['blogs'], response_model=List[schemas.BlogSchema], status_code=status.HTTP_202_ACCEPTED)
+def show_all_blogs(db: Session= Depends(get_db)):
     blogs = db.query(models.Blog).options(joinedload(models.Blog.owner)).all()
     return jsonable_encoder(blogs)
 
 
-@app.get('/blog/{id}', tags=['blogs'], response_model=schemas.ShowBlog, status_code=status.HTTP_202_ACCEPTED)
-def show(id: int, db: Session = Depends(get_db), get_current_user:schemas.User = Depends(oauth2.get_current_user)):
+@app.get('/blog/{id}', tags=['blogs'], response_model=schemas.BlogSchema, status_code=status.HTTP_202_ACCEPTED)
+def show(id: int, db: Session = Depends(get_db), 
+         get_current_user:schemas.User = Depends(oauth2.get_current_user),
+         ):
     blog = db.query(models.Blog).options(joinedload(models.Blog.owner)).filter(models.Blog.id == id).first()
     if not blog:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail= f'Blog with id {id} is not available')
+    
     return jsonable_encoder(blog)
 
 
 @app.patch('/blogs/edit/{blog_id}', tags=['blogs'], status_code=status.HTTP_202_ACCEPTED)
-def edit_blog(blog_id:int, request:schemas.ShowBlog, db:Session=Depends(get_db), get_current_user:schemas.User = Depends(oauth2.get_current_user)):
+def edit_blog(blog_id:int, request:schemas.BlogSchema, 
+              db:Session=Depends(get_db), 
+              get_current_user:schemas.User = Depends(oauth2.get_current_user),
+              current_user: schemas.User = Depends(oauth2.get_current_user)
+              ):
     blog = db.query(models.Blog).filter(models.Blog.id == blog_id).first()
     if not blog:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'Blog with id {id} is not found')
     
+    if blog.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to edit this blog")
+
     blog.title = request.title or blog.title
     blog.body = request.body or blog.body
     db.commit()
@@ -74,10 +87,18 @@ def edit_blog(blog_id:int, request:schemas.ShowBlog, db:Session=Depends(get_db),
 
     
 @app.delete('/blogs/delete/{blog_id}', tags=['blogs'], status_code=status.HTTP_202_ACCEPTED)
-def delete_blog(blog_id:int, db:Session=Depends(get_db), get_current_user:schemas.User = Depends(oauth2.get_current_user)):
+def delete_blog(blog_id:int, 
+                db:Session=Depends(get_db), 
+                get_current_user:schemas.User = Depends(oauth2.get_current_user),
+                current_user: schemas.User = Depends(oauth2.get_current_user)
+                ):
     blog = db.query(models.Blog).filter(models.Blog.id == blog_id).first()
     if not blog:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'Blog with id {id} is not found')
+    
+    if blog.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to edit this blog")
+
     db.delete(blog)
     db.commit()
     return {'detail': f'Blog with id {id} deleted successfully'}
